Collapse interfaces.
