+++ airbyte-integrations/connectors/destination-clickhouse/src/main/kotlin/io/airbyte/integrations/destination/clickhouse/config/ClickhouseDirectLoadDatabaseInitialStatusGatherer.kt
@@ -4,17 +4,17 @@
 
 package io.airbyte.integrations.destination.clickhouse.config
 
-import io.airbyte.cdk.load.client.AirbyteClient
+import io.airbyte.cdk.load.client.TableOperationsClient
 import io.airbyte.cdk.load.orchestration.db.BaseDirectLoadInitialStatusGatherer
 import io.airbyte.cdk.load.orchestration.db.TempTableNameGenerator
 import jakarta.inject.Singleton
 
 @Singleton
 class ClickhouseDirectLoadDatabaseInitialStatusGatherer(
-    airbyteClient: AirbyteClient,
+    tableOperationsClient: TableOperationsClient,
     tempTableNameGenerator: TempTableNameGenerator,
 ) :
     BaseDirectLoadInitialStatusGatherer(
-        airbyteClient,
+        tableOperationsClient,
         tempTableNameGenerator,
     )
